+++ tests/dep-time-travel.sh
@@ -34,3 +34,11 @@ if [[ "$TRAVIS_NODE_VERSION" -lt "8" ]]; then
   echo "Downgrading eslint-import-resolver-typescript..."
   npm i --no-save eslint-import-resolver-typescript@1.0.2
 fi
+
+if [ "${ESLINT_VERSION}" = '8' ]; then
+  # This is a workaround for the crash in the initial processing of the ESLint class.
+  echo "Installing self"
+  npm i --no-save eslint-plugin-import@'.' -f
+  echo "Build self"
+  npm run build
+fi